requirements.txt
gradio==5.9.1
langchain-openai==0.2.14
langchain==0.3.13
langchain-community==0.3.13
python-dotenv==1.0.0